refactor(core): defer and separate out score compounding
Refactor the scoring to not increase chips and multiplier on the fly and instead defer the compounding of the total chips and multiplier and the final score computation to the last possible point in time. Move those final calculations into a separate file. This is done in preparation of introducing a layer that can handle math with really, really big numbers with plain JavaScript cannot (which makes the calculator unable to deal with really, really high scores).