feat(ui): add card count input for easier same-card duplication
Add a count number input to both joker and playing cards in the UI. Increasing the count acts exactly like duplicating the card as many times as the count was increased. This makes it easier to input hands that hast a lot of copies of the exact same card (e.g. hundreds of steel, red-sealed Kings held in hand).

Resolves #11.