feat(ui): use input+datalist instead of select for better search experience
Replace all instances of `<select>` with `<input>` and `<datalist>` for a better type-ahead search experience. Typing "moon" will now match "Shoot the Moon" and "To the Moon" (for example) in the joker control when before this wouldn't yield any matches. This comes with the downside that arbitrary values can be typed into the input. To cover for this problem, some form validation logic was added which should make it somewhat obvious that the current input is incomplete/invalid.